requirements.txt
great-expectations==1.3.12
pandas>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
anthropic>=0.5.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
sqlalchemy>=2.0.23
openpyxl>=3.1.2
python-multipart>=0.0.6 